+++ src/ConnectorRuntimeConfig.ts
@@ -1,5 +1,6 @@
 import { RuntimeConfig } from "@nmshd/runtime";
 import * as log4js from "log4js";
+import { ConnectorRuntimeModuleConfiguration } from "./ConnectorRuntimeModule";
 
 export interface ConnectorRuntimeConfig extends RuntimeConfig {
     database: {
@@ -8,4 +9,14 @@ export interface ConnectorRuntimeConfig extends RuntimeConfig {
     };
 
     logging: log4js.Configuration;
+
+    modules: Record<string, ConnectorRuntimeModuleConfiguration>;
+
+    infrastructure: {
+        httpServer: {
+            enabled: boolean;
+            apiKey: string;
+            cors?: any;
+        };
+    };
 }