+++ src/ConnectorRuntimeModule.ts
@@ -2,7 +2,7 @@ import { ModuleConfiguration, RuntimeModule } from "@nmshd/runtime";
 import { ConnectorRuntime } from "./ConnectorRuntime";
 
 export interface ConnectorRuntimeModuleConfiguration extends ModuleConfiguration {
-    someConnectorOnlyProperty: string;
+    requiredInfrastructure?: string[];
 }
 
 export abstract class ConnectorRuntimeModule<TConfig extends ConnectorRuntimeModuleConfiguration = ConnectorRuntimeModuleConfiguration> extends RuntimeModule<