+++ src/infrastructure/httpServer/HttpServer.ts
@@ -8,9 +8,8 @@ import http from "http";
 import { Server } from "typescript-rest";
 import TypescriptRestIOC from "typescript-rest-ioc";
 import { buildInformation } from "../../buildInformation";
-import { ConnectorRuntimeModule, ConnectorRuntimeModuleConfiguration } from "../../ConnectorRuntimeModule";
-import { Envelope } from "../coreHttpApi/common/Envelope";
-import { HttpError } from "../coreHttpApi/common/HttpError";
+import { ConnectorInfrastructure, InfrastructureConfiguration } from "../ConnectorInfastructure";
+import { Envelope, HttpError } from "./common";
 import { HttpMethod } from "./HttpMethod";
 import { csrfErrorHandler } from "./middlewares/csrfErrorHandler";
 import { genericErrorHandler, RouteNotFoundError } from "./middlewares/genericErrorHandler";
@@ -37,14 +36,13 @@ export interface ControllerConfig {
     baseDirectory: string;
 }
 
-export interface HttpServerModuleConfiguration extends ConnectorRuntimeModuleConfiguration {
-    port: number;
-    hostname: string;
+export interface HttpServerConfiguration extends InfrastructureConfiguration {
+    port?: number;
     apiKey: string;
     cors?: CorsOptions;
 }
 
-export default class HttpServerModule extends ConnectorRuntimeModule<HttpServerModuleConfiguration> {
+export class HttpServer extends ConnectorInfrastructure<HttpServerConfiguration> {
     private app: Application;
     private readonly customEndpoints: CustomEndpoint[] = [];
     private readonly controllers: ControllerConfig[] = [];
@@ -219,8 +217,9 @@ export default class HttpServerModule extends ConnectorRuntimeModule<HttpServerM
     public async start(): Promise<void> {
         this.configure();
         return await new Promise((resolve) => {
-            this.server = this.app.listen(this.configuration.port, () => {
-                this.logger.info(`Listening on port ${this.configuration.port}`);
+            const port = this.configuration.port ?? 80;
+            this.server = this.app.listen(port, () => {
+                this.logger.info(`Listening on port ${port}`);
                 resolve();
             });
         });