+++ src/infrastructure/httpServer/common/HttpError.ts
@@ -1,3 +1,5 @@
+import { DocumentationLink } from "../../../DocumentationLink";
+
 export class HttpError {
     public readonly code: string;
     public readonly message: string;
@@ -10,7 +12,8 @@ export class HttpError {
         this.time = new Date().toISOString();
         this.code = code;
         this.message = message;
-        this.docs = `https://enmeshed.eu/integrate/error-codes#${code}`;
+
+        this.docs = DocumentationLink.integrate__errorCodes(code);
     }
 
     public static forDev(code: string, message: string, stacktrace: string[], details: string): HttpErrorDev {