+++ src/modules/coreHttpApi/controllers/FilesController.ts
@@ -2,8 +2,8 @@ import { OwnerRestriction, TransportServices } from "@nmshd/runtime";
 import express from "express";
 import { Inject } from "typescript-ioc";
 import { Accept, Context, ContextAccept, ContextResponse, FileParam, FormParam, GET, Path, PathParam, POST, Return, ServiceContext } from "typescript-rest";
+import { Envelope } from "../../../infrastructure";
 import { BaseController, Mimetype } from "../common/BaseController";
-import { Envelope } from "../common/Envelope";
 
 @Path("/api/v1/Files")
 export class FilesController extends BaseController {